move reify errors into the primaryDomainCommunicator
